// mobile/src/types/api.ts
export interface LoginPayload {
  email: string
  password: string
}

export interface RegisterPayload {
  email: string
  password: string
  password_confirm: string
}

export interface TokenResponse {
  access_token: string
  token_type: string
  expires_in: number
}